--- templates/repo/find/files.tmpl
@@ -1,21 +1,0 @@
-{{template "base/head" .}}
-<div role="main" aria-label="{{.Title}}" class="page-content repository">
-	{{template "repo/header" .}}
-	<div class="ui container">
-		<div class="tw-flex tw-items-center">
-			<a href="{{$.RepoLink}}">{{.RepoName}}</a>
-			<span class="tw-mx-2">/</span>
-			<div class="ui input tw-flex-1">
-				<input id="repo-file-find-input" type="text" autofocus data-url-data-link="{{.DataLink}}" data-url-tree-link="{{.TreeLink}}">
-			</div>
-		</div>
-		<table id="repo-find-file-table" class="ui single line fixed table">
-			<tbody>
-			</tbody>
-		</table>
-		<div id="repo-find-file-no-result" class="ui row center tw-mt-8 tw-hidden">
-			<h3>{{ctx.Locale.Tr "repo.find_file.no_matching"}}</h3>
-		</div>
-	</div>
-</div>
-{{template "base/footer" .}}
