+++ templates/repo/view_file_tree.tmpl
@@ -8,7 +8,7 @@
 </div>
 
 {{/* TODO: Dynamically move components such as refSelector and createPR here */}}
-<div id="view-file-tree" class="tw-overflow-y-auto tw-overflow-x-visible tw-h-full is-loading"
+<div id="view-file-tree" class="tw-overflow-auto tw-h-full is-loading"
 	data-repo-link="{{.RepoLink}}"
 	data-tree-path="{{$.TreePath}}"
 	data-current-ref-name-sub-url="{{.RefTypeNameSubURL}}"
