+++ web_src/css/themes/theme-gitea-dark.css
@@ -244,6 +244,7 @@ gitea-theme-meta-info {
   --color-highlight-fg: #87651e;
   --color-highlight-bg: #352c1c;
   --color-overlay-backdrop: #080808c0;
+  --color-danger: var(--color-red);
   accent-color: var(--color-accent);
   color-scheme: dark;
 }
