+++ web_src/js/features/repo-findfile.ts
@@ -1,13 +1,7 @@
-import {svg} from '../svg.ts';
-import {toggleElem} from '../utils/dom.ts';
-import {pathEscapeSegments} from '../utils/url.ts';
-import {GET} from '../modules/fetch.ts';
+import {createApp} from 'vue';
+import RepoFileSearch from '../components/RepoFileSearch.vue';
 
 const threshold = 50;
-let files: Array<string> = [];
-let repoFindFileInput: HTMLInputElement;
-let repoFindFileTableBody: HTMLElement;
-let repoFindFileNoResult: HTMLElement;
 
 // return the case-insensitive sub-match result as an array:  [unmatched, matched, unmatched, matched, ...]
 // res[even] is unmatched, res[odd] is matched, see unit tests for examples
@@ -73,48 +67,14 @@ export function filterRepoFilesWeighted(files: Array<string>, filter: string) {
   return filterResult;
 }
 
-function filterRepoFiles(filter: string) {
-  const treeLink = repoFindFileInput.getAttribute('data-url-tree-link');
-  repoFindFileTableBody.innerHTML = '';
-
-  const filterResult = filterRepoFilesWeighted(files, filter);
-
-  toggleElem(repoFindFileNoResult, !filterResult.length);
-  for (const r of filterResult) {
-    const row = document.createElement('tr');
-    const cell = document.createElement('td');
-    const a = document.createElement('a');
-    a.setAttribute('href', `${treeLink}/${pathEscapeSegments(r.matchResult.join(''))}`);
-    a.innerHTML = svg('octicon-file', 16, 'tw-mr-2');
-    row.append(cell);
-    cell.append(a);
-    for (const [index, part] of r.matchResult.entries()) {
-      const span = document.createElement('span');
-      // safely escape by using textContent
-      span.textContent = part;
-      span.title = span.textContent;
-      // if the target file path is "abc/xyz", to search "bx", then the matchResult is ['a', 'b', 'c/', 'x', 'yz']
-      // the matchResult[odd] is matched and highlighted to red.
-      if (index % 2 === 1) span.classList.add('ui', 'text', 'red');
-      a.append(span);
-    }
-    repoFindFileTableBody.append(row);
-  }
-}
-
-async function loadRepoFiles() {
-  const response = await GET(repoFindFileInput.getAttribute('data-url-data-link'));
-  files = await response.json();
-  filterRepoFiles(repoFindFileInput.value);
-}
-
 export function initFindFileInRepo() {
-  repoFindFileInput = document.querySelector('#repo-file-find-input');
-  if (!repoFindFileInput) return;
-
-  repoFindFileTableBody = document.querySelector('#repo-find-file-table tbody');
-  repoFindFileNoResult = document.querySelector('#repo-find-file-no-result');
-  repoFindFileInput.addEventListener('input', () => filterRepoFiles(repoFindFileInput.value));
-
-  loadRepoFiles();
+  const fileSearchContainer = document.querySelector('.repo-file-search-container');
+  if (!fileSearchContainer) return;
+  createApp(RepoFileSearch, {
+    repoLink: fileSearchContainer.getAttribute('data-repo-link'),
+    currentRefNameSubURL: fileSearchContainer.getAttribute('data-current-ref-name-sub-url'),
+    treeListUrl: fileSearchContainer.getAttribute('data-tree-list-url'),
+    noResultsText: fileSearchContainer.getAttribute('data-no-results-text'),
+    placeholder: fileSearchContainer.getAttribute('data-placeholder'),
+  }).mount(fileSearchContainer);
 }
